Fix comparison: don't use 'is' for string comparison
"is" and "is not" compare addresses of objects while == and != compare values of objects.
The problem is that some immutable objects are cached/singletons, but some are not.
There is only one "None" object, anything compared to None should be "is None" or "is not None".
But strings could have multiple copies, but not necessarily, and the same is true for integers. 

Long story short: the ```is not "auto"``` and the ```is "float32"``` could work, but not guaranteed, and should be replaced with `!=` and `==`.